+++ class/App/Model.php
@@ -3,8 +3,23 @@ namespace App;
 
 abstract class Model extends \R\Model
 {
-    
-    public function _app(){
+
+    public function __construct($id)
+    {
+        parent::__construct($id);
+        $c = new \ReflectionClass(get_class($this));
+        if ($static = $c->getStaticProperties()) {
+            foreach ($static as $k => $v) {
+                if ($r = self::_sv("$c->name/$k")) {
+                    $c->setStaticPropertyValue($k, $r);
+                }
+            }
+        }
+
+    }
+
+    public function _app()
+    {
         return \App::_();
     }
 
@@ -167,18 +182,23 @@ abstract class Model extends \R\Model
             }
         }
 
-        if($static=$c->getStaticProperties()){
+        if ($static = $c->getStaticProperties()) {
             $decamlize = function ($string) {
                 return strtolower(preg_replace(['/([a-z\d])([A-Z])/', '/([^_])([A-Z][a-z])/'], '$1_$2', $string));
             };
             $field = $decamlize($function);
 
-            $f="_".$field;
+            $f = $field;
+            if (array_key_exists($f, $static)) {
+                return $static[$f][$this->$field];
+            }
+
+            $f = "_" . $field;
             if (array_key_exists($f, $static)) {
                 return $static[$f][$this->$field];
             }
         }
-        
+
         $f = "_" . $function;
 
         if (property_exists($class, $f)) {
@@ -188,14 +208,16 @@ abstract class Model extends \R\Model
         return parent::__call($function, $args);
     }
 
-    public static function Query(){
-        $q= new Query(get_called_class());
+    public static function Query()
+    {
+        $q = new Query(get_called_class());
         $q->select();
         $q->from(self::_table()->name);
         return $q;
     }
 
-    public static function _sv($name){
+    public static function _sv($name)
+    {
         return self::_app()->sv($name);
     }